Создаю модуль util.js и помещаю в него функцию генерирующую сл число
