+++ index.html
@@ -235,7 +235,6 @@
     </section>
   </template>
   <script src="js/main.js" type="module"></script>
-  <script src="js/functions.js"></script>
 </body>
 
 </html>
